refactor: rename query.ts to read.ts
