+++ lib/actions/load-edit-page.ts
@@ -1,6 +1,6 @@
 'use server';
 
-import { getRecipes } from '@/lib/repository/recipe-store/query';
+import { getRecipes } from '@/lib/repository/recipe-store/read';
 import { createWhereIdClause } from '@/lib/repository/recipe-store/utils';
 import { recipePageSchema } from '@/lib/translation/schema';
 import { convertPageToForm } from '@/lib/translation/parsers';
